refactor playlist-loader, isolate m3u8 static parsing, fragment and level-key class
